feat: add parsemangalist, queryparameters and review changes
